+++ src/vectors/vectors_properties.c
@@ -1,13 +1,13 @@
 #include "vectors.h"
 
-double	vlength2(t_vector3 v)
+float	vlength2(t_vector3 v)
 {
 	if (DEBUG && v.w == POINT)
 		printf("Warning: Getting the squared length of a point\n");
 	return (v.x * v.x + v.y * v.y + v.z * v.z);
 }
 
-double	vlength(t_vector3 v)
+float	vlength(t_vector3 v)
 {
 	if (DEBUG && v.w == POINT)
 		printf("Warning: Getting the length of a point\n");
